Fig. 5. EurekAlert! Articles With or Without DOIs (2018)
This dataset begins in the year 2009 and ends at 2018. The total number of articles on to EurekAlert! in 2009 is 8,250. Using the query: 
 
SELECT*FROM scraped_data_all WHERE DATE >=’2009-01-01' AND DATE <=’2009-12-31' AND DOI IS NULL; 
 
yields the result of 8,250. Thus, no articles during this time have a DOI. Using the same query: 
 
SELECT*FROM scraped_data_all WHERE DATE >=’2018-01-01' AND DATE <=’2018-12-31' AND DOI IS NULL; 
 

We see that 12,635 articles have no DOI. Again, using: 
 
SELECT CASE WHEN DOI <> ‘/N’ THEN ‘/Y’ 
		ELSE DOI 
	END AS DOI_exists 
FROM scraped_data_all 
WHERE DATE >=2018-01-01' 
AND DATE <=’2018-09-30' 
AND DOI IS NOT NULL 
 
This yields the result of 12,042 that do have a DOI in 2018. The data shows that in the beginning, EurekAlert! articles had no DOIs, but as more articles were added, more articles with DOIs were included on the site. In 2018, 51% of articles had no DOI, while 49% did (Fig. 5.)